Initial commit of project source
The current source does not yet work correctly, as the offset
is incorrect and therefore the code is not restored at the right
address.  This is due to function starting addresses as defined
in an elf file are not the same as function addresses derived
at runtime; I believe the function address retrieved at runtime
is after the function prologue.  However, further testing is
needed.