Fixed restoration code
Now the code is correctly restored to the correct addresses.

The problem was not with the function prologue (the address given
was in fact correct), but rather that the offset I had calculated
was incorrect and gave the offset of the text section from the
start of the file instead of the offset of the init function from
the start of the text section.