Finished prototype decryption library functions
Now, the host application simply can call elide_restore to
restore all functions; the library automatically determines
whether to decrypt the bytes or not based on the secret metadata.

However, right now, only the code for decrypting an on-disk
secret file is implemented, so right now there is no way to have
the "non-encrypted" type of secret (which would be sent from the
server over an encrypted connection).

In addition, this has yet to be merged with the server part, so
we currently get the metadata (which should be sent over a secure
connection from the server) from the file on disk.  This is a
simplification for testing that will change once we merge with the
client/server code that handles getting remote resources.